manifests, and accompanying code changes, to bring up bridge as RHDH sidecar containers
So this was interesting out of the gate because RHDH forces the backstage deployment to
`automountServiceAccountToken: false`
meaning we have to manually mount the SA credentials/certs we want the bridge to use.

Ultimately, I could got the correct k8s json patch specification for the entire deployment such that
it would add/merge the the new secret `rhdh-rhoai-bridge-token` volume with all the other volumes mounted
as part of the ai-rhdh-installer into the deployment.

Instead, I import the secret via `envFrom` into our containers, and also add a volumeMount to the root volume
declared for the deployment (i.e. /opt/app-root-src) and then store the ca.crt from the env var into a file which
I can then seed into the REST config used to establish k8s client communications.

I also dynamically set the pod's IP and namespace as env vars as well.  This allows us to create our storage
configmap as well as supply the root IP for the communication between the 4 containers in the pod.

Added a 10.*:9090 allowed host for the sidecar location service.

Access to the backstage backend did not work using pod IP:port nor service IP; have to continue to use the OCP route's address.

Lastly, so far I did not need to create additional Services for the bridge's REST endpoints to prove the config map is populated
when a new model is created/deployed from the model registry.  But as we complete integrating with our new backstage extensions,
if need be that would need to be added as new instances in our assets/sidecar directory, using the same selector as the
backstage service.